The zarr library

The Zarr library provides an OCaml implementation of the Zarr version 3 storage format specification for chunked & compressed multi-dimensional arrays, designed for use in parallel computing. The storage format is used by many companies including Google, NASA, Microsoft and many others. Zarr's goal is to provide the following features:

Chunk multi-dimensional arrays along any dimension.

Store arrays in memory, on disk, inside a Zip file or any remote storage backend.

Read and write arrays concurrently from multiple threads or processes.

Organize arrays into hierarchies using groups.

See Zarr V3 specification.

Author: Zolisa Bleki

Entry Point

The entry point of this library is the module Zarr.